Added chat commands
 - /a allows staff and server owners to make announcements. Affected lines: 255, 317-332
 - /pop echoes the number of people on the server. Affected lines: 256, 333-335

Quick note about /a: staff and private server owners had the ability to negate the warning text that came after the name by doing `/view your text here &#0;` and warning the non-existent user.  This would make the game do this: `bls1999 has given your text here`.  I assume this will be broken with htmlspecialchars, so instead, server owners and staff now have the option to use the /a command to craft custom messages.  This can be useful for tournaments and general fun-having.