Add option to recount number of people in a guild
Some guilds' member count is way off (i.e. Lost and Found, PR2 Staff, etc).  A good fix for this would be to allow admins to recount the number of members in the guild using the number of rows for members in the database.

This needs edits from you to be fully functional.  Please enable $db->call('guild_update', array()) to be able to update the owner_id and the member_count.