Key changes to first artifact finder mechanic
This pull contains some key changes to the first artifact finder mechanic.  The script will now:

 - Echo a prize box that says "Bubble Set" and has the bubble body as an icon
 - There's too much going on when the user finds the artifact first.  Instead of popping up a congratulatory message, PM it to the user (check my code on line 45 to make sure I'm using `$db->call` correctly)
 - Award the prize via gain_part() (taken from Player.php) and auto-set the prize upon return to the lobby

Of course, check my code, but this is a lot simpler than the code was before since it uses functions that already exist.  Plus, it actually works. :)

LMK if/when you merge this code and get it running on Isabel so I can test if it works for awarding it to the first finder.