Add more error text
I think I may have solved it.  Not sure if `SELECT COUNT(*)` can have `LIMIT 0, 1` as an argument.  I removed it so we'll see what happens.  I do know that it's an issue with that function though because that's the only one that doesn't have an error explicitly defined (until this commit).

If that wasn't right, I'll figure out what the heck is wrong with this eventually.