Update vault_descriptions.php
Another optional change. I noticed that there are 2 versions of the private server image: one with a transparent background and the other with a blue one. The transparent one is currently used for both the 1 day and 30 day options, so this makes it so the 30 day private server has the blue one and doesn't have to share.