Add time left command in Player.php
This is a command for owners of Private Servers. When they type /timeleft or /ps it will tell them when their private server will expire (I think). This is my first ever time doing something in php but I believe this should not cause any errors. The only issue I think that could occur is how this expire time might be shown. From what I've seen I think it is in a date format but I am not 100% sure. Anyone can modify this if they want but if you can tell me why you are changing what you did that will help. Thanks!