add $suppress_error to user not found as well
I'm using it in update_account.php to check if a name already exists before changing it to that one.  The easiest way that I can see is to try to grab a user ID for the name.  If it fails, then that means the script can continue changing the name.