if this doesn't fix kong badges idk
Analyzed the sent information to the server... it seems that when Guests log in, they don't send the correct $login->user_name, and don't send a $login->user_pass.  I think that may be what's been holding us back unless, for some reason, you coded into the game that if you don't enter a name and password into the boxes in the client that it just shouldn't award badges. That's a very specific code though... I have good faith that this will work.  Knock wood.