actually fix users_select_top
While debugging, I noticed the script was requiring GROUP BY with the aggregate function SUM() being used.  When I grouped by name, it told me pr2.rank needed to be included since it was in the SUM() function.  It subsequently continued to tell me I needed more columns until I ended up with the monstrosity that is line 14.  From there, it works perfectly, although I'm not sure if that's an effective method of fixing the problem.

An alternative fix is to disable ONLY_FULL_GROUP_BY in the database.  More info here: https://stackoverflow.com/questions/23921117/disable-only-full-group-by