Fix FAH pt. 79
Oh, you know. Folding at Home decided to update their fast teampage URL for the first time since PR2 came out, so it broke yet again.

I'd like to be able to go a year without this breaking, even if it's not my fault in the first place. That would be nice.