include ' in htmlspecialchars
Small fix to the urlify function so links and display text that involve the &#039; character (apos) will be escaped properly. I'm proclaiming this "the ENT_QUOTES band-aid." This will need to be corrected throughout the rest of the repo, so once these few pending pulls of mine are merged, I'll get started on it (would be too many merge conflicts if I started now).