fix no pr2/epart data when awarding prize
Allow prizes to be awarded to people who don't have entries in the pr2 or epic_upgrades tables in the db. This shouldn't be an issue too often, since most of the contests that award prizes award them to active users; however, for a contest like CotM which can award prizes to inactive level makers, this will stop an error from showing.