Artifact fairness, ch. 2
In an effort to make things fairer for all, artifacts are now announced ahead of time. That way, people can race to be the first and know exactly where to look, instead of some having advantages over others.

By virtue of the artifact update, this also implements a level of the week system into PR2. More details to follow.

(Also fixed a really niche and precise potential exploit w/ the artifact awards in Game.php)